mvp/python-api/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.2.0
appnope==0.1.3
asttokens==2.4.1
attrs==23.2.0
backoff==2.2.1
beautifulsoup4==4.12.2
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
comm==0.2.1
contourpy==1.2.0
cryptography==41.0.7
cycler==0.12.1
dataclasses-json==0.6.3
debugpy==1.8.0
decorator==5.1.1
Deprecated==1.2.14
deprecation==2.1.0
distro==1.9.0
effdet==0.4.1
emoji==2.9.0
executing==2.0.1
filelock==3.13.1
filetype==1.2.0
flatbuffers==23.5.26
fonttools==4.47.0
frozenlist==1.4.1
fsspec==2023.12.2
gotrue==2.1.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
huggingface-hub==0.20.1
humanfriendly==10.0
idna==3.6
iopath==0.1.10
ipykernel==6.28.0
ipython==8.19.0
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
jupyter_client==8.6.0
jupyter_core==5.7.0
kiwisolver==1.4.5
langchain==0.0.354
langchain-community==0.0.8
langchain-core==0.1.6
langdetect==1.0.9
langsmith==0.0.77
layoutparser==0.3.4
lxml==5.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.2
matplotlib-inline==0.1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.8
networkx==3.2.1
nltk==3.8.1
numpy==1.26.3
omegaconf==2.3.0
onnx==1.15.0
onnxruntime==1.15.1
openai==1.6.1
opencv-python==4.9.0.80
packaging==23.2
pandas==2.1.4
parso==0.8.3
pdf2image==1.16.3
pdfminer.six==20221105
pdfplumber==0.10.3
pexpect==4.9.0
pikepdf==8.11.2
pillow==10.2.0
platformdirs==4.1.0
portalocker==2.8.2
postgrest==0.13.1
prompt-toolkit==3.0.43
protobuf==4.25.1
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pycocotools==2.0.7
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.17.2
pyparsing==3.1.1
pypdf==3.17.4
pypdfium2==4.25.0
pytesseract==0.3.10
python-dateutil==2.8.2
python-docx==1.1.0
python-dotenv==1.0.0
python-iso639==2024.1.2
python-magic==0.4.27
python-multipart==0.0.6
python-pptx==0.6.23
pytube==15.0.0
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
rapidfuzz==3.6.1
realtime==1.0.2
regex==2023.12.25
requests==2.31.0
safetensors==0.4.1
scipy==1.11.4
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.25
stack-data==0.6.3
storage3==0.7.0
StrEnum==0.4.15
supabase==2.3.1
supafunc==0.3.3
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tiktoken==0.5.2
timm==0.9.12
tokenizers==0.15.0
torch==2.1.2
torchvision==0.16.2
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
transformers==4.36.2
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
unstructured==0.11.8
unstructured-client==0.15.1
unstructured-inference==0.7.18
unstructured.pytesseract==0.3.12
urllib3==2.1.0
wcwidth==0.2.12
websockets==11.0.3
wrapt==1.16.0
XlsxWriter==3.1.9
yarl==1.9.4
youtube-transcript-api==0.6.2
